Refactoring to align old code to the new styling standard  used.

+++ src/rendering/swapchain.h
@@ -15,7 +15,7 @@
 
 namespace gigno {
 
-	const int MAX_FRAMES_IN_FLIGHT = 2;
+	const size_t MAX_FRAMES_IN_FLIGHT = 2;
 
 	/*
 	Per-Entity data pushed to the shader.
